+++ tests/tests/neg/surface/reifyFnPtr00.rs
